simpleclaw-backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
yookassa==3.10.0
paramiko==3.5.0
requests==2.32.3
google-auth==2.37.0
